Supprime la duplication de code de mapping
Le code avait été dupliqué lorsque nous avons séparé les routes publiques et privées dans des fichiers distincts.
